pipeline-scripts/requirements.txt
asn1crypto==0.24.0
azure-common==1.1.23
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.5
azure-functions==1.0.0b5
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-common==1.4.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
idna==2.8
pycparser==2.19
python-dateutil==2.8.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3